+++ FunctionRevisionProject/task07.cpp
@@ -4,7 +4,7 @@
 #include <iostream>
 using namespace std;
 
-long long subtract(long long a, int b);
+long long subtract(long long a, short b);
 
 int main()
 {
@@ -12,7 +12,7 @@ int main()
     return 0;
 }
 
-long long subtract(long long a, int b)
+long long subtract(long long a, short b)
 {
     return a - b;
 }
